+++ src/Gmail_App.py
@@ -1,7 +1,18 @@
-from Google import Create_Service
+from datetime import datetime
+import pytz
 from googleapiclient.errors import HttpError
 
-class Gmail_App:
+from Google import Create_Service
+
+def audit_log(type, string):
+        date_format='%m/%d/%Y %H:%M:%S %Z'
+        date = datetime.now(tz=pytz.utc)
+        date = date.astimezone(pytz.timezone('US/Pacific'))
+        fd = open("logfile.txt", 'a')
+        fd.writelines(f'{date.strftime(date_format)} | {type} | {string}\n')
+        fd.close()
+
+class GmailApp:
     def __init__(self, first, last, email, label_name):
         self.first = first
         self.last = last
@@ -20,14 +31,14 @@ class Gmail_App:
         try:
             return self.service.users().messages().list(userId=self.email, labelIds='INBOX').execute()
         except HttpError as error:
-            print(f'An error occurred: {error}')        # SELF NOTE WRITE TO LOGGING
+            audit_log("Error", f'An error occurred: {error}')
             return False
     
     def get_message(self, message_id):
         try:
             return self.service.users().messages().get(userId=self.email, id=message_id, format='metadata').execute()
         except HttpError as error:
-            print(f'An error occurred: {error}')        # SELF NOTE WRITE TO LOGGING
+            audit_log("Error", f'An error occurred: {error}')
             return False
 
     def add_quarantine(self, message_ids):
@@ -39,7 +50,7 @@ class Gmail_App:
             }
             self.service.users().messages().batchModify(userId=self.email, body=body).execute()
         except HttpError as error:
-            print(f'An error occurred: {error}')        # SELF NOTE WRITE TO LOGGING
+            audit_log("Error", f'An error occurred: {error}')
 
 
     def check_label(self, name):
@@ -52,4 +63,4 @@ class Gmail_App:
                 label = self.service.users().labels().create(userId=self.email,body={"name":name}).execute()
                 return label.get("id")
         except HttpError as error:
-            print(f'An error occurred: {error}')        # SELF NOTE WRITE TO LOGGING
+            audit_log("Error", f'An error occurred: {error}')
