+++ src/app.py
@@ -1,38 +1,32 @@
-from Gmail_App import Gmail_App
-from datetime import datetime
-import pytz
 import re
 
-def audit_log(type, string):
-        date_format='%m/%d/%Y %H:%M:%S %Z'
-        date = datetime.now(tz=pytz.utc)
-        date = date.astimezone(pytz.timezone('US/Pacific'))
-        fd = open("logfile.txt", 'a')
-        fd.writelines(f'{date.strftime(date_format)} | {type} | {string}\n')
-        fd.close()
+from Gmail_App import GmailApp, audit_log
 
 def handler(app):
         black_lists = open('blacklist.txt', 'r')
-        bcontent = black_lists.read()
+        blocked_emails = black_lists.readlines()
         emails = app.list_inbox()
         blacklist_ids = []
-        for messages in emails['messages']:
+        for messages in emails.get('messages'):
             m = app.get_message(messages.get('id'))
             headers = (m.get("payload")).get("headers")
             subject = next((header.get("value") for header in headers if header["name"] == "Subject"), None)
             sender = next((header.get("value") for header in headers if header["name"] == "From"), None)
-            sender_email = re.search('<(.*)>', sender)
-            if sender_email.group(1)==bcontent.strip():
-                blacklist_ids.append(messages.get('id'))
-                audit_log("Filter", f'{sender} : {subject}')
-            # else:
-            #     implicit deny
-            #     print("placeholder")
+            sender_email = (re.search('<(.*)>', sender)).group(1)
+            for blocked_email in blocked_emails:
+                if sender_email == blocked_email.strip():
+                    blacklist_ids.append(messages.get('id'))
+                    audit_log("Filter", f'{sender} : {subject}')
+                    break
+                # else:
+                #     implicit deny
+                #     print("placeholder")
                 
 
         if blacklist_ids:
             app.add_quarantine(blacklist_ids)
 
-app = Gmail_App("Test", "Account", "spamfilterbottester@gmail.com", "Quarantine")
-handler(app)
-
+if __name__ == "__main__":
+    app = GmailApp("Test", "Account", "spamfilterbottester@gmail.com", "Quarantine")
+    handler(app)
+    print("Done")
